requirements.txt
streamlit
openai
azure-search-documents>=11.4.0
azure-core>=1.24.0
beautifulsoup4>=4.11.0
requests>=2.28.0
uuid